backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.7
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
fpdf2>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
